Minor cleanup & good news
It seems like our memory management  knowledge was enough.
The only mem-leaks that exist are from libfreetype (or some other lib I do not recall the name of) which did exist before and are harmless.

I have been running conky for quite a few hours now and its memory footprint is small; it *IS* growing but will not pose a problem unless you do not restart your OS for something like a year.